+++ src/types/theme/BaseFontSize.ts
@@ -7,10 +7,5 @@ export type BaseFontSize = {
     md: string
     lg: string
     xl: string
-    '2xl': string
-    '3xl': string
-    '4xl': string
-    '5xl': string
-    '6xl': string
 }