+++ src/types/theme/BaseFontWeight.ts
@@ -2,12 +2,10 @@
  * Font weight scale
  */
 export type BaseFontWeight = {
-    thin: number
     light: number
     regular: number
     medium: number
     semibold: number
     bold: number
-    extrabold: number
 }