+++ src/types/theme/BaseLineHeight.ts
@@ -4,9 +4,7 @@
 export type BaseLineHeight = {
     none: number
     tight: number
-    snug: number
     normal: number
     relaxed: number
-    loose: number
 }