+++ src/types/theme/BaseRadius.ts
@@ -3,11 +3,11 @@
  */
 export type BaseRadius = {
     none: string
+    xs: string
     sm: string
     md: string
     lg: string
     xl: string
-    '2xl': string
     full: string
 }