+++ src/types/theme/BaseShadows.ts
@@ -8,7 +8,5 @@ export type BaseShadows = {
     md: string
     lg: string
     xl: string
-    '2xl': string
-    inner: string
 }