+++ src/types/theme/BaseSpacing.ts
@@ -4,14 +4,10 @@
  */
 export type BaseSpacing = {
     none: string
-    px: string
     xs: string
     sm: string
     md: string
     lg: string
     xl: string
-    '2xl': string
-    '3xl': string
-    '4xl': string
 }