+++ src/types/theme/BaseTransition.ts
@@ -5,6 +5,5 @@ export type BaseTransition = {
     fast: string
     normal: string
     slow: string
-    slower: string
 }